feat: colorful lifecycle callbacks
